fix: Move some TPP tables to a table core
As these are intended to be shared by other backends in future.
